:recycle: refactor Refactoring coupon Entity and order Entity
