+++ src/main/kotlin/com/highv/ecommerce/domain/order_details/enumClass/ComplainStatus.kt
@@ -1,0 +1,16 @@
+package com.highv.ecommerce.domain.order_details.enumClass
+
+
+enum class ComplainStatus {
+
+    NONE,
+
+    REFUND_REQUESTED,
+    REFUNDED,
+    REFUND_REJECTED,
+
+    EXCHANGE_REQUESTED,
+    EXCHANGED,
+    EXCHANGE_REJECTED,
+
+}
