+++ src/main/kotlin/com/highv/ecommerce/domain/order_details/enumClass/OrderStatus.kt
@@ -1,0 +1,6 @@
+package com.highv.ecommerce.domain.order_details.enumClass
+
+enum class OrderStatus {
+    EXCHANGE,
+    REFUND,
+}
