+++ src/main/kotlin/com/highv/ecommerce/domain/order_master/controller/ProductsOrderController.kt
@@ -1,0 +1,42 @@
+package com.highv.ecommerce.domain.order_master.controller
+
+import com.highv.ecommerce.common.dto.DefaultResponse
+import com.highv.ecommerce.domain.order_master.dto.CouponRequest
+import com.highv.ecommerce.domain.order_master.dto.OrderStatusRequest
+import com.highv.ecommerce.domain.order_master.service.ProductsOrderService
+import com.highv.ecommerce.infra.security.UserPrincipal
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import org.springframework.security.access.prepost.PreAuthorize
+import org.springframework.security.core.annotation.AuthenticationPrincipal
+import org.springframework.web.bind.annotation.*
+
+@RestController
+@RequestMapping("/api/v1/")
+class ProductsOrderController(
+    private val productsOrderService: ProductsOrderService
+) {
+
+    @PreAuthorize("hasRole('BUYER')")
+    @PostMapping("/payments")
+    fun requestPayment(
+        @AuthenticationPrincipal userPrincipal: UserPrincipal,
+        @RequestBody couponRequest: CouponRequest
+    ): ResponseEntity<DefaultResponse>
+        = ResponseEntity.status(HttpStatus.OK).body(productsOrderService.requestPayment(userPrincipal.id, couponRequest))
+
+
+    @PreAuthorize("hasRole('SELLER')")
+    @PatchMapping("/order_status/{orderId}")
+    fun updateOrderStatus(
+        @PathVariable("orderId") orderId: Long,
+        @RequestBody orderStatusRequest: OrderStatusRequest,
+        @AuthenticationPrincipal userPrincipal: UserPrincipal,
+    ): ResponseEntity<DefaultResponse>
+        = ResponseEntity.status(HttpStatus.OK).body(productsOrderService.updateOrderStatus(orderId, orderStatusRequest, userPrincipal.id))
+
+
+
+
+
+}
