+++ src/main/kotlin/com/highv/ecommerce/domain/order_master/dto/CouponRequest.kt
@@ -1,0 +1,5 @@
+package com.highv.ecommerce.domain.order_master.dto
+
+data class CouponRequest(
+    val couponIdList: List<Long>
+)
