+++ src/main/kotlin/com/highv/ecommerce/domain/order_master/dto/OrderStatusRequest.kt
@@ -1,0 +1,7 @@
+package com.highv.ecommerce.domain.order_master.dto
+
+import com.highv.ecommerce.domain.order_master.enumClass.StatusCode
+
+data class OrderStatusRequest(
+    val statusCode: StatusCode
+)
