+++ src/main/kotlin/com/highv/ecommerce/domain/order_master/enumClass/StatusCode.kt
@@ -1,0 +1,18 @@
+package com.highv.ecommerce.domain.order_master.enumClass
+
+enum class StatusCode {
+                      
+    PENDING,
+
+    CANCELLED,
+
+    ORDERED,
+
+    PAYMENTED,
+
+    DELIVERED_PREPARING,
+
+    SHIPPING,
+
+    DELIVERED,
+}
