+++ src/main/kotlin/com/highv/ecommerce/domain/order_master/repository/ProductsOrderJpaRepository.kt
@@ -1,0 +1,8 @@
+package com.highv.ecommerce.domain.order_master.repository
+
+import com.highv.ecommerce.domain.order_master.entity.OrderMaster
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface ProductsOrderJpaRepository: JpaRepository<OrderMaster, Long> {
+
+}
