--- src/main/kotlin/com/highv/ecommerce/domain/order_status/repository/OrderStatusRepository.kt
@@ -1,20 +1,0 @@
-package com.highv.ecommerce.domain.order_status.repository
-
-import com.highv.ecommerce.domain.order_status.entity.OrderStatus
-
-interface OrderStatusRepository {
-
-    fun findAllByShopIdAndProductsOrderId(shopId: Long, productsOrderId: Long): List<OrderStatus>
-
-    fun findAllByBuyerId(id: Long): List<OrderStatus>
-
-    fun save(orderStatus: OrderStatus): OrderStatus
-
-    fun saveAll(orderStatuses: List<OrderStatus>): List<OrderStatus>
-
-    fun findAllByShopId(shopId: Long): List<OrderStatus>
-
-    fun findByItemCartIdAndBuyerId(orderStatusId: Long, buyerId: Long): OrderStatus?
-
-    fun findByIdAndShopId(orderStatusId: Long, shopId: Long): OrderStatus?
-}
