--- src/main/kotlin/com/highv/ecommerce/domain/products_order/dto/CouponRequest.kt
@@ -1,5 +1,0 @@
-package com.highv.ecommerce.domain.products_order.dto
-
-data class CouponRequest(
-    val couponIdList: List<Long>
-)
