--- src/main/kotlin/com/highv/ecommerce/domain/products_order/dto/OrderStatusRequest.kt
@@ -1,7 +1,0 @@
-package com.highv.ecommerce.domain.products_order.dto
-
-import com.highv.ecommerce.domain.products_order.enumClass.StatusCode
-
-data class OrderStatusRequest(
-    val statusCode: StatusCode
-)
