--- src/main/kotlin/com/highv/ecommerce/domain/products_order/enumClass/OrderStatusType.kt
@@ -1,6 +1,0 @@
-package com.highv.ecommerce.domain.products_order.enumClass
-
-enum class OrderStatusType {
-    EXCHANGE,
-    REFUND,
-}
