--- src/main/kotlin/com/highv/ecommerce/domain/products_order/enumClass/StatusCode.kt
@@ -1,18 +1,0 @@
-package com.highv.ecommerce.domain.products_order.enumClass
-
-enum class StatusCode {
-                      
-    PENDING,
-
-    CANCELLED,
-
-    ORDERED,
-
-    PAYMENTED,
-
-    DELIVERED_PREPARING,
-
-    SHIPPING,
-
-    DELIVERED,
-}
