--- src/main/kotlin/com/highv/ecommerce/domain/products_order/repository/ProductsOrderJpaRepository.kt
@@ -1,8 +1,0 @@
-package com.highv.ecommerce.domain.products_order.repository
-
-import com.highv.ecommerce.domain.products_order.entity.ProductsOrder
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface ProductsOrderJpaRepository: JpaRepository<ProductsOrder, Long> {
-
-}
